+++ src/hooks/useCouponApplicable.ts
@@ -1,7 +1,7 @@
 import { useRecoilValue } from 'recoil';
-import { cartOrderTotalPriceSelector } from '../../recoil/CartItem/selectors/selectors';
-import { isCouponExpired, isCouponMinimumAmount, isCouponAvaliable } from '../../utils/couponValidator';
-import { Coupon } from '../../types/Coupon.type';
+import { cartOrderTotalPriceSelector } from '../recoil/CartItem/selectors/selectors';
+import { isCouponExpired, isCouponMinimumAmount, isCouponAvaliable } from '../utils/couponValidator';
+import { Coupon } from '../types/Coupon.type';
 
 const useCouponApplicable = () => {
   const totalPrice = useRecoilValue(cartOrderTotalPriceSelector);
